mipsevm: rename package, make available as library instead of main package
