+++ mipsevm/patch.go
@@ -1,4 +1,4 @@
-package main
+package mipsevm
 
 import (
 	"bytes"
